Feature/resources

+++ app/src/main/java/com/uitopic/restockmobile/features/resources/domain/models/Batch.kt
@@ -6,6 +6,6 @@ data class Batch(
     val id: String,
     val userId: Int?,
     val customSupply: CustomSupply?,
-    val stock: Int,
+    val stock: Double,
     val expirationDate: String?
 )
